tests/wpt/tests/html/canvas/element/canvas-host/2d.canvas.host.size.attributes.default.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.canvas.host.size.attributes.default</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.canvas.host.size.attributes.default</h1>
<p class="desc">Default width/height when attributes are missing</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Default width/height when attributes are missing");
_addTest(function(canvas, ctx) {

  _assertSame(canvas.width, 100, "canvas.width", "100");
  _assertSame(canvas.height, 50, "canvas.height", "50");

});
</script>
